Cleanup serialization
Remove unnecessary deriving from Serialize
Use string methods for serialization where
it makes sense
