+++ components/common/src/templating/package.rs
@@ -10,8 +10,7 @@ use crate::hcore::{fs,
                         users},
                    package::{PackageIdent,
                              PackageInstall},
-                   util::{deserialize_using_from_str,
-                          serialize_using_to_string}};
+                   util::serde_string};
 use serde::{ser::SerializeStruct,
             Serialize,
             Serializer};
@@ -62,10 +61,9 @@ impl Env {
     }
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Deserialize)]
 pub struct Pkg {
-    #[serde(deserialize_with = "deserialize_using_from_str",
-            serialize_with = "serialize_using_to_string")]
+    #[serde(with = "serde_string")]
     pub ident: PackageIdent,
     pub origin: String,
     pub name: String,
