+++ components/core/src/package/target.rs
@@ -423,7 +423,7 @@ impl<'d> serde::Deserialize<'d> for PackageTarget {
     fn deserialize<D>(deserializer: D) -> result::Result<Self, D::Error>
         where D: serde::Deserializer<'d>
     {
-        util::deserialize_using_from_str(deserializer)
+        util::serde_string::deserialize(deserializer)
     }
 }
