+++ components/sup/src/manager/service/mod.rs
@@ -118,7 +118,7 @@ enum BindStatus<'a> {
     Unknown(Error),
 }
 
-#[derive(Debug, Serialize)]
+#[derive(Debug)]
 pub struct Service {
     pub service_group:       ServiceGroup,
     pub bldr_url:            String,
@@ -134,7 +134,6 @@ pub struct Service {
     pub initialized:         bool,
     pub user_config_updated: bool,
 
-    #[serde(skip_serializing)]
     config_renderer: CfgRenderer,
     // Note: This field is really only needed for serializing a
     // Service in the gateway (see ServiceProxy's Serialize
@@ -173,27 +172,22 @@ pub struct Service {
     /// We don't serialize because this is purely runtime information
     /// that should be reconciled against the current state of the
     /// census.
-    #[serde(skip_serializing)]
     unsatisfied_binds: HashSet<ServiceBind>,
     hooks: HookTable,
     config_from: Option<PathBuf>,
     manager_fs_cfg: Arc<FsCfg>,
-    #[serde(rename = "process")]
     supervisor: Arc<Mutex<Supervisor>>,
     svc_encrypted_password: Option<String>,
     health_check_interval: HealthCheckInterval,
 
-    #[serde(skip_serializing)]
     /// Whether a service's default configuration changed on a package
     /// update. Used to control when templates are re-rendered.
     defaults_updated: bool,
-    #[serde(skip_serializing)]
     gateway_state: Arc<RwLock<GatewayState>>,
 
     /// A "handle" to the never-ending future that periodically runs
     /// health checks on this service. This is the means by which we
     /// can stop that future.
-    #[serde(skip_serializing)]
     health_check_handle: Option<sup_futures::FutureHandle>,
 }
